#KE-12525 event log roll up

+++ core/src/main/scala/org/apache/spark/scheduler/EventLoggingListener.scala
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
+// scalastyle:off
 package org.apache.spark.scheduler
 
 import java.io._
@@ -22,15 +22,14 @@ import java.net.URI
 import java.nio.charset.StandardCharsets
 import java.util.Locale
 
+import org.apache.commons.compress.utils.CountingOutputStream
 import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
-
 import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.{FileSystem, FSDataOutputStream, Path}
+import org.apache.hadoop.fs.{FSDataOutputStream, FileSystem, Path}
 import org.apache.hadoop.fs.permission.FsPermission
 import org.json4s.JsonAST.JValue
 import org.json4s.jackson.JsonMethods._
-
 import org.apache.spark.{SPARK_VERSION, SparkConf}
 import org.apache.spark.deploy.SparkHadoopUtil
 import org.apache.spark.internal.Logging
@@ -69,6 +68,12 @@ private[spark] class EventLoggingListener(
   private val testing = sparkConf.get(EVENT_LOG_TESTING)
   private val outputBufferSize = sparkConf.get(EVENT_LOG_OUTPUT_BUFFER_SIZE).toInt
   private val fileSystem = Utils.getHadoopFileSystem(logBaseDir, hadoopConf)
+
+  private val ROLL_LOG_DIR_NAME_PREFIX = "eventlog_v2_"
+  private val ROLL_LOG_FILE_NAME_PREFIX = "events_"
+  private var rollIndex: Long = 0L
+  private val eventRollFileMaxLength = sparkConf.get(EVENT_LOG_ROLLING_MAX_FILE_SIZE)
+
   private val compressionCodec =
     if (shouldCompress) {
       Some(CompressionCodec.createCodec(sparkConf))
@@ -88,7 +93,13 @@ private[spark] class EventLoggingListener(
   private[scheduler] val loggedEvents = new ArrayBuffer[JValue]
 
   // Visible for tests only.
-  private[scheduler] val logPath = getLogPath(logBaseDir, appId, appAttemptId, compressionCodecName)
+  private[scheduler] var logPath = getLogPath(logBaseDir, appId, appAttemptId, compressionCodecName)
+
+  // For roll event log
+  private[scheduler] var rollDirPath = new Path(new Path(logBaseDir),ROLL_LOG_DIR_NAME_PREFIX+nameForAppAndAttempt(appId,appAttemptId))
+  private var currentRollLogFilePath: Path = _
+  private var countingRollOutputStream: Option[CountingOutputStream] = None
+
 
   /**
    * Creates the log file in the configured log directory.
@@ -98,34 +109,48 @@ private[spark] class EventLoggingListener(
       throw new IllegalArgumentException(s"Log directory $logBaseDir is not a directory.")
     }
 
-    val workingPath = logPath + IN_PROGRESS
-    val path = new Path(workingPath)
-    val uri = path.toUri
-    val defaultFs = FileSystem.getDefaultUri(hadoopConf).getScheme
-    val isDefaultLocal = defaultFs == null || defaultFs == "file"
+    if (!sparkConf.get(EVENT_LOG_ENABLE_ROLLING)) {
+      val path = new Path(logPath + IN_PROGRESS)
+      initLogFile(path){ os => new PrintWriter(os) }
+    } else {
+      if (fileSystem.exists(rollDirPath) && shouldOverwrite) {
+        fileSystem.delete(rollDirPath, true)
+      }
+      fileSystem.mkdirs(rollDirPath, LOG_FILE_PERMISSIONS)
+      rollEventLogFile()
+    }
 
+  }
+
+
+  protected def initLogFile(path: Path)(fnSetupWriter: OutputStream => PrintWriter): Unit = {
     if (shouldOverwrite && fileSystem.delete(path, true)) {
       logWarning(s"Event log $path already exists. Overwriting...")
     }
 
-    /* The Hadoop LocalFileSystem (r1.0.4) has known issues with syncing (HADOOP-7844).
-     * Therefore, for local files, use FileOutputStream instead. */
+    val defaultFs = FileSystem.getDefaultUri(hadoopConf).getScheme
+    val isDefaultLocal = defaultFs == null || defaultFs == "file"
+    val uri = path.toUri
+
+    // The Hadoop LocalFileSystem (r1.0.4) has known issues with syncing (HADOOP-7844).
+    // Therefore, for local files, use FileOutputStream instead.
     val dstream =
-      if ((isDefaultLocal && uri.getScheme == null) || uri.getScheme == "file") {
-        new FileOutputStream(uri.getPath)
-      } else {
-        hadoopDataStream = Some(fileSystem.create(path))
-        hadoopDataStream.get
-      }
+    if ((isDefaultLocal && uri.getScheme == null) || uri.getScheme == "file") {
+      new FileOutputStream(uri.getPath)
+    } else {
+      hadoopDataStream = Some(
+        SparkHadoopUtil.createFile(fileSystem, path, sparkConf.get(EVENT_LOG_ALLOW_EC)))
+      hadoopDataStream.get
+    }
 
     try {
-      val cstream = compressionCodec.map(_.compressedOutputStream(dstream)).getOrElse(dstream)
+      val cstream = compressionCodec.map(_.compressedOutputStream(dstream))
+        .getOrElse(dstream)
       val bstream = new BufferedOutputStream(cstream, outputBufferSize)
-
-      EventLoggingListener.initEventLog(bstream, testing, loggedEvents)
       fileSystem.setPermission(path, LOG_FILE_PERMISSIONS)
-      writer = Some(new PrintWriter(bstream))
-      logInfo("Logging events to %s".format(logPath))
+      EventLoggingListener.initEventLog(bstream, testing, loggedEvents)
+      logInfo(s"Logging events to $path")
+      writer = Some(fnSetupWriter(bstream))
     } catch {
       case e: Exception =>
         dstream.close()
@@ -133,8 +158,53 @@ private[spark] class EventLoggingListener(
     }
   }
 
+  def rollEventLogFile(): Unit ={
+    closeWriter()
+    rollIndex += 1
+    val now = System.currentTimeMillis()
+    val base = s"${ROLL_LOG_FILE_NAME_PREFIX}${rollIndex}_" + nameForAppAndAttempt(appId, appAttemptId) +
+      "_" + now
+    if( currentRollLogFilePath != null){
+      fileSystem.rename(currentRollLogFilePath,new Path(currentRollLogFilePath.toUri.getPath + "_" + now))
+    }
+    currentRollLogFilePath = new Path(rollDirPath, base)
+
+    initLogFile(currentRollLogFilePath){ os =>
+      countingRollOutputStream = Some(new CountingOutputStream(os))
+      new PrintWriter(
+        new OutputStreamWriter(countingRollOutputStream.get, StandardCharsets.UTF_8))
+    }
+  }
+
+  def nameForAppAndAttempt(appId: String, appAttemptId: Option[String]): String = {
+    val base = sanitizeDirName(appId)
+    if (appAttemptId.isDefined) {
+      base + "_" + sanitizeDirName(appAttemptId.get)
+    } else {
+      base
+    }
+  }
+
+  def sanitizeDirName(str: String): String = {
+    str.replaceAll("[ :/]", "-").replaceAll("[.${}'\"]", "_").toLowerCase(Locale.ROOT)
+  }
+
+  protected def closeWriter(): Unit = {
+    writer.foreach(_.close())
+  }
+
   /** Log the event as JSON. */
   private def logEvent(event: SparkListenerEvent, flushLogger: Boolean = false) {
+
+    if(sparkConf.get(EVENT_LOG_ENABLE_ROLLING)){
+      writer.foreach { w =>
+        val currentLen = countingRollOutputStream.get.getBytesWritten
+        if (currentLen  > eventRollFileMaxLength) {
+          rollEventLogFile()
+        }
+      }
+    }
+
     val eventJson = JsonProtocol.sparkEventToJson(event)
     // scalastyle:off println
     writer.foreach(_.println(compact(render(eventJson))))
@@ -244,24 +314,31 @@ private[spark] class EventLoggingListener(
   def stop(): Unit = {
     writer.foreach(_.close())
 
-    val target = new Path(logPath)
-    if (fileSystem.exists(target)) {
-      if (shouldOverwrite) {
-        logWarning(s"Event log $target already exists. Overwriting...")
-        if (!fileSystem.delete(target, true)) {
-          logWarning(s"Error deleting $target")
+    if(!sparkConf.get(EVENT_LOG_ENABLE_ROLLING)){
+      val target = new Path(logPath)
+      if (fileSystem.exists(target)) {
+        if (shouldOverwrite) {
+          logWarning(s"Event log $target already exists. Overwriting...")
+          if (!fileSystem.delete(target, true)) {
+            logWarning(s"Error deleting $target")
+          }
+        } else {
+          throw new IOException("Target log file already exists (%s)".format(logPath))
         }
-      } else {
-        throw new IOException("Target log file already exists (%s)".format(logPath))
       }
-    }
-    fileSystem.rename(new Path(logPath + IN_PROGRESS), target)
-    // touch file to ensure modtime is current across those filesystems where rename()
-    // does not set it, -and which support setTimes(); it's a no-op on most object stores
-    try {
-      fileSystem.setTimes(target, System.currentTimeMillis(), -1)
-    } catch {
-      case e: Exception => logDebug(s"failed to set time of $target", e)
+      fileSystem.rename(new Path(logPath + IN_PROGRESS), target)
+      // touch file to ensure modtime is current across those filesystems where rename()
+      // does not set it, -and which support setTimes(); it's a no-op on most object stores
+      try {
+        fileSystem.setTimes(target, System.currentTimeMillis(), -1)
+      } catch {
+        case e: Exception => logDebug(s"failed to set time of $target", e)
+      }
+    }else {
+      if( currentRollLogFilePath != null){
+        fileSystem.rename(currentRollLogFilePath,new Path(currentRollLogFilePath.toUri.getPath + "_" + System.currentTimeMillis()))
+        currentRollLogFilePath = null
+      }
     }
   }
